backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
pyyaml==6.0.2
langchain-groq==0.2.4
langchain-core==0.3.29
apscheduler==3.10.4
sqlalchemy==2.0.36
python-telegram-bot==21.6
aiofiles==24.1.0
